Moved DEMCollection and dDEM to xdem/dem.py
