+++ xdem/__init__.py
@@ -1,1 +1,3 @@
-from . import coreg, dem, examples, spatial_tools, volume, spstats
+from xdem.dem import DEM, DEMCollection, dDEM
+
+from . import coreg, dem, examples, spatial_tools, spstats, volume
